+++ CHANGELOG.md
@@ -12,6 +12,11 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added `metrics` property to `<manifold-connection>` to opt-in for metric collection (#724)
 - Added performance monitoring of GraphQL endpoints (#726)
 
+### Removed
+
+- Data Dog logging removed (#729)
+- Removed `manifold-performance` component (#729)
+
 ## [0.6.5] - 2019-11-13
 
 ### Added